done 2 tasks
done for 4.  add validation check for each field (use logic) and 6.  A notification and history for organizer and guest on status change or incoming new applicant

but heven't do 1. In the create event, the event start date must be > 3 days (based on today). Else ask the user to change to another start date.
7. Make sure the publish later date < event date
in 4.  add validation check for each field (use logic)